requirements.txt
PyQt5==5.13.1
python-chess==0.28.3
